refactor: move hooks and rename its files
