+++ src/pages/executive/CheckInScan.tsx
@@ -22,12 +22,12 @@ import DirectInputFab from "components/DirectInputFab";
 import ResultChip, { ResultChipRefs } from "components/ResultChip";
 import ResultPopup, { ResultPopupRefs } from "components/ResultPopup";
 import { useTitleSet } from "libs/title";
-import { useAspidaClient, useAuthState } from "libs/auth/useAuth";
-import { useRequirePermission } from "libs/auth/useRequirePermission";
-import useErrorHandler from "libs/useErrorHandler";
+import { useAspidaClient, useAuthState } from "hooks/auth/useAuth";
+import { useRequirePermission } from "hooks/auth/useRequirePermission";
+import useErrorHandler from "hooks/useErrorHandler";
 import { getStringDateTimeBrief, getStringTime } from "libs/stringDate";
-import { useWristBandPaletteColor } from "libs/wristBandColor";
-import useReset from "libs/useReset";
+import useWristBandPaletteColor from "hooks/useWristBandColor";
+import useReset from "hooks/useReset";
 import { StatusColor } from "types/statusColor";
 import api, { Reservation, Term } from "@afes-website/docs";
 import clsx from "clsx";
