+++ src/pages/executive/RegisterSpare.tsx
@@ -20,9 +20,9 @@ import DirectInputFab from "components/DirectInputFab";
 import ResultChip, { ResultChipRefs } from "components/ResultChip";
 import ResultPopup, { ResultPopupRefs } from "components/ResultPopup";
 import { useTitleSet } from "libs/title";
-import { useAspidaClient, useAuthState } from "libs/auth/useAuth";
-import { useRequirePermission } from "libs/auth/useRequirePermission";
-import useErrorHandler from "libs/useErrorHandler";
+import { useAspidaClient, useAuthState } from "hooks/auth/useAuth";
+import { useRequirePermission } from "hooks/auth/useRequirePermission";
+import useErrorHandler from "hooks/useErrorHandler";
 import { StatusColor } from "types/statusColor";
 import api from "@afes-website/docs";
 import clsx from "clsx";
